Decomposed the large AG Grid config file
