+++ out.log
@@ -3039,3 +3039,7 @@
 {"message":"\n> start\n> cross-env NODE_ENV=production node .\n\n","timestamp":"14-07 02:52:19 -07:00","type":"out","process_id":1,"app_name":"veud"}
 {"message":"","timestamp":"14-07 17:37:02 -07:00","type":"out","process_id":0,"app_name":"veud"}
 {"message":"","timestamp":"14-07 17:37:02 -07:00","type":"out","process_id":1,"app_name":"veud"}
+{"message":"\n> start\n> cross-env NODE_ENV=production node .\n\n","timestamp":"14-07 18:57:59 -07:00","type":"out","process_id":0,"app_name":"veud"}
+{"message":"\n> start\n> cross-env NODE_ENV=production node .\n\n","timestamp":"14-07 18:57:59 -07:00","type":"out","process_id":1,"app_name":"veud"}
+{"message":"","timestamp":"14-07 22:38:43 -07:00","type":"out","process_id":0,"app_name":"veud"}
+{"message":"","timestamp":"14-07 22:38:43 -07:00","type":"out","process_id":1,"app_name":"veud"}
