+++ tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "include": ["**/*.ts", "**/*.tsx", "app/routes/lists+/index.tsx", "app/utils/lists/thumbnail-renderer.tsx", "app/routes/lists+/$watchlist_grid.update-watchlist.jsx", "app/components/search-add-watchlist-entry.jsx", "app/utils/lists/column-functions.jsx"],
+  "include": ["**/*.ts", "**/*.tsx"],
   "compilerOptions": {
     "lib": ["DOM", "DOM.Iterable", "ES2022"],
     "isolatedModules": true,
